+++ examples/local/index.ts
@@ -1,7 +1,7 @@
-import { runSwarmLoop } from '../run';
+import { runDemoLoop } from '../run';
 import { localAgent } from './localAgent';
 
-runSwarmLoop({
+runDemoLoop({
   initialAgentMessage: 'Would you like to know how many days are in a month?',
   initialAgent: localAgent,
 });