+++ examples/pokemon/index.ts
@@ -1,8 +1,8 @@
 import 'dotenv/config';
 import { pokemonAgent } from './pokemonAgent';
-import { runSwarmLoop } from '../run';
+import { runDemoLoop } from '../run';
 
-runSwarmLoop({
+runDemoLoop({
   initialAgentMessage: 'What Pokémon would you like to know about?',
   initialAgent: pokemonAgent,
 });